[FIX] AlignedFlowLayout: incorrect arrangement of items fixed
The issue:
Incorrect arrangement of items across the available horizontal space
when the user interface is initially rendered, and the browser window is
resized up.

The root cause of the issue:
The number of spacers were calculated during rendering time based on the
width of the root element of the document
(document.documentElement.clientWidth).
However, this value may change if the browser is resized up.

The solution:
Increase the number of spacers.

Change-Id: I0a9fc7886625c082a047f277a75ce4b0d31be666