[FIX] FlexibleColumnLayout: Not showing arrows in Demo App is fixed
Problem: arrows are not showing at all in Demo App
(Master-Detail), because _adaptableContentChange event is not
fired from NavContainer, due to the fact that its DOM reference
is still missing, when addPage method is executed.

Solution: Introduce new event - _onNavContainerRendered, which is
fired when onAfterRendering of the NavContainer is executed for the
very first time. Thus, we can guarantee that at this moment, there is
some content on the page and the FlexibleColumnLayout's arrows are not
shown on a blank page.

Change-Id: I2577a1b4fe08d6f5dbdf4dbdbf8f3a86d6c67786
BCP: 1880353153