[FIX] Link: tabindex attribute for disabled links removed
Removed tabindex attribute for disabled links, due to jQuery
specifics. For jQuery, an anchor is :focusable if it has either
href or tabindex attribute. It doesn't consider tabindex's value
however, causing a link without href and tabindex="-1" to be
regarded as :focusable.

tabindex was removed only for disabled links, because for enabled
links:

   - tabindex="0" is needed when href doesn't exist. Otherwise
     our link won't be focusable due to standard HTML anchor behavior
   - tabindex="-1" is needed when there is no text. Otherwise the
     link would be focusable, causing strange readings. For more info,
     see incident 1670557122

Change-Id: Ie6707ebf623faf7753fd7fe69d449b7231116a80
BCP: 1870038330