[INTERNAL] sap.ui.unified.Currency: improvement of the demokit sample
The sap.ui.unified.Currency controls from the "Currency" named example
are shown in the corresponding three-letter code and are now displayed
in a sap.m.List controls. According to the guidelines,
the sap.ui.unified.Currency control should be displayed in sap.m.List,
sap.m.Table or in sap.ui.layout.form.Form control and it is recommended
to always show the corresponding three-letter code.

The example "Currency in Table" is updated in order to be alligned to the
gudelines.

JIRA: BGSOFUIBALKAN-4398

Change-Id: Ifaa32bf20507c65d8bb3480e4ff768b925291c14