[FIX] Configuration: Ensure gregorian calendar for legacy patterns 1-6
We correctly set the calendar type for Japanese, Islamic and Persian
calendar if the respective legacy pattern is selected, but we didn't set
it to Gregorian for legacy pattern 1 to 6. This was causing the default
calendar of the locale to be used with the pattern for gregorian (e.g.
ar_SA, th_TH).

Change-Id: Iaf2dbc2d77926917393ddd5dcd6b231f449f5015