[FIX] ObjectPageHeaderActionButton: RTA "Rename" action is fixed
The Problem: "sap.uxap.ObjectPageHeaderActionButton" rename RTA action is
available in icon buttons.

The Root Cause: "sap.uxap.ObjectPageHeaderActionButton" inherits
"sap.m.Button". This causes the default behavior to be as it is in the
parent class.
The default behavior is that when we have a text property in a button,
the rename action is always available. If we have an icon button with no
text property, the rename is disabled.
In the case of "sap.uxap.ObjectPageHeaderActionButton" we can have both,
the text property and the icon property. This is because even if we have
icon only button, we still need the text, since it's used as a description,
when the "sap.uxap.ObjectPageHeaderActionButton" overflows.

The Solution: We check if the button has an icon property and if it has,
we disable the "rename" action. Even if it has "text" property.

Change-Id: I62b8fb4bc7ea319ba66c451a3a4f3919beb8f845
BCP: 1880354063