[FIX] sap.m.TimePickerSliders: collapsing before animation complete is now working.
Previously, if the given slider is collapsed (#setIsExpanded(false)),
before slider animation is completed, then this used to potentially call
TimePickerSlider.prototype._scrollerSnapped function with snapped index = -1.

The reason was that the snapped index was not yet calculated, due to the
not completed animation.

If this is the first animation, the snap index remains its initial value
of -1, which is not accepted by the function _scrollerSnapped.

Change-Id: Ie55f8c4bc2e308b695153e599848d39d9fb79477
BCP: 1880363483