[FIX] XMLComposite: Event propagation in cloning
Note:
In contrast to other controls where cloning means that the events from 
outside like e.g the surrounding controller of a view need not to be 
adapted as they are inherited this is not the case for an XMLComposite.

This control comprises of a templated fragment where the innre controls 
use the composite as controller (event handler).
When cloning these contol now the event handler points to the template 
which is not true. It must the the clone composite. 

With the first draft to fix that problem of the event cloning we 
interchanged the template with in the only in the head content, but not
in the inner content. This leaded to problmes as the content is a
fragment with single aggregation content.

Imagine now an HBox containing a button and an input. The button 
influences the input that points to the value of the control.
With the old coding we did not access the button.

Change-Id: I3a1a00c6d1e50574cd9e02120fa87a661c8a0955