[FIX] OverflowToolbar: Additional LayoutData check is implemented
Problem: Error in console appears when FlexItemData is used, instead of
ToolbarLayoutData/OverflowToolbarLayoutData.
The issue appears due to the fact that there is no check for the type
of the LayoutData and getShrinkable method is executed directly on it.

Solution: Add additional check for the type of the LayoutData before calling
its specific methods.

Change-Id: Ib18cff1228b59d5cc5072c5180609bf4c08f20fa
BCP: 1880387343