[FIX] sap/ui/dom/computedStylePolyfill: document.body is not there yet
The situation is the same as before. Firefox still returns null for getComputedStyle
when called in an iframe with style "display:none".

The previous fix failed when it was called synchronously from the HEAD element,
as no document.body is present at that point in time.

The fake body approach is also used by jQuery itself.

In the incident the getComputedStyle call is triggered by the sap.m library.js for the
font-size check.

CR-Id: 002075125900002569602018
Change-Id: I6c6ca21714e8b5eb5849527460098ce3084e5df6
BCP: 002028376500003153742018
(cherry picked from commit afe894e1db2a172b1828481465f67009c28c93e2)