[FIX] Core: Trigger ThemeCheck when a library has been initialized
The ThemeCheck should always be triggered when
sap.ui.getCore().initLibrary is called, not only when it is loaded via
sap.ui.getCore().loadLibrary/loadLibraries.

This might happen when a control is required before the corresponding
library has been loaded (via loadLibrary).
In this case there was no "ThemeChanged" event fired after the library
is loaded which results into wrong display / behavior of controls which
listen to that event.

The sap/m/OverflowToolbar test is affected by this changed as sap/m/Menu
is loaded which has a dependency to sap/ui/unified/Menu and in this case
the theme check is now triggered. This causes a timing issue. The test
has been solved by just loading sap.ui.unified initially as it's
required for the test anyway.

Change-Id: Ibe3453b7aeadb714594ac186348756db589afc39
BCP: 1870314182
BCP: 1770568826