[FIX] sap.uxap.ObjectPageLayout: Non existing sub-section is now handled
The Problem: When we search for the closest visible sub section, if
there is none, the execution of the code stops with exception.

The Solution: We double check if a sub-section is existing, and if it
isn't, we set the closest ID to be the one of the section.

Change-Id: I9f44e7482701839b3697e223614a7a3daeb9bb73
BCP: 1870038109
(cherry picked from commit 99512e2d88f38d50b9dfd07ccaeaec86f406b4b2)
Fixes: https://github.com/SAP/openui5/issues/2174
CR-Id: 002075125800003172882018