[FIX] sap.uxap.ObjectPageLayout: Non existing sub-section is now handled
The Problem: When we search for the closest visible sub section, if
there is none, the execution of the code stops with exception.

The Solution: We double check if a sub-section is existing, and if it
isn't, we set the closest ID to be the one of the section.

BCP: 1970039639
(cherry picked from commit 99512e2d88f38d50b9dfd07ccaeaec86f406b4b2)

CR-Id: 002075125800000071392019
(cherry picked from commit 0a2c78f5a62a43f0e55a26642efcbbe866ed09a6)

Change-Id: I5016586f4aa6f3b94252c710987837d2f8a75b39