[FIX] sap.uxap.ObjectPageLayout: Non existing sub-section is now handled
The Problem: When we search for the closest visible sub section, if
there is none, the execution of the code stops with exception.

The Solution: We double check if a sub-section is existing, and if it
isn't, we set the closest ID to be the one of the section.

Change-Id: I9f44e7482701839b3697e223614a7a3daeb9bb73
BCP: 1870038109