[FIX] FlexibleColumnLayout: Fixed suspend and resume resize handlers
The problem:
On application with FlexibleColumnLayout with routing and DynamicPage
control when you navigate to a mid section (or back to begin section)
containing DynamicPage which was hidden the DynamicPage control is not
visible on the screen and you have empty section of the
FlexibleColumnLayout

The root cause:
Current logic in code upon navigation suspended ResizeHandlers for all
3 sections but restored it only for the last section as the internal
_iResumeResizeHandlerTimeout was recorded for the FCL control instead
of creating one for every section. Thus the _iResumeResizeHandlerTimeout
was cleared for the begin and mid section and was only active for
section end. This caused the resume of the suspended resize handlers
to never happen for column begin and mid

The solution:
Implement _iResumeResizeHandlerTimeout per section so every section has
a resume resize handlers scheduled

Change-Id: I4b7a481987016bdc04929f40c07b48e5fabed2c1
BCP: 1880392685
BCP: 1880391328