[INTERNAL] v2.ODataModel: Canceling requests
With this change the ODataModel V2 request header 'sap-cancel-on-close'
is defaulted to 'false'. Only requests triggered by ODataModel.read and
security token requests are send with 'sap-cancel-on-close' equals true by
default.

If numbers are drawn at the back-end, the previous behavior could lead to inconsistent
data, since some numbers are lost through canceling. A typical
scenario for canceling a requests is closing the browser window.

Furthermore, the header can now be set for all requests by passing the header via parameter to
the ODataModel constructor, or make use of the setHeader API, or for an single request as header parameter
to the ODataModel APIs (read, update, ...).

Change-Id: Ifdb62cc33f9eeafcaf42f4d5122bd9e7c561ec56
CR-Id: 002075125900002784912018
BCP: 1880400449