[INTERNAL] sap.m.ViewSettingsDialog: improvement of the demokit sample
The "Table - View Settings & Context Menu" example
in the demokit for the sap.m.ViewSettingsDialog is improved.
The icon buttons for sorting, filtering and grouping
are placed now directly in the toolbar and with each,
opens the sap.m.ViewSettingsDialog with just the corresponding tab,
as recommended in the Fiori guidelines.

JIRA: BGSOFUIBALKAN-4412

Change-Id: Ie580fd415ece6f915651eb08178b8a28bc597400