[INTERNAL][FIX] ui5loader-autoconfig: Detect jquery-ui-position
jQuery.prototype.position is provided by jQuery itself and is extended
by jquery-ui-position. Therefore this check is always true which leads
to an issue when sap-ui-core-nojQuery.js is used and only jquery.js is
loaded beforehand but not jquery-ui-position.js

jQuery.ui.position is added by jQuery UI Position, which is a better
detection.

Change-Id: Ib74079e2836f2d25d97335c46326a114c7d97f7a