[INTERNAL] grunt docs: better support for AMD module exports in api.json
The api.json should not only list classes, interfaces, enums and
typedefs as top-level entities, but also any direct, public export of
AMD modules.

The longname of a symbol is checked for the pattern

  module:some/module/name

Members of an export should not become top-level entities and are
filtered out. The filtering is based on the assumption that they have a
compound name like

  module:some/module/name.member
  module:some/module/name#member
  module:some/module/name~member

Note: when the characters '#', '~' or '.' are to be used in module
names, a more complex filtering has to be implemented.

Change-Id: Ib6655d43f3b40850b9fd47ce0232ef8ce27c052d