[INTERNAL] jquery.sap.encoder: Remove unnecessary lazy getter
The idea of lazy loading the sanitizeHTML module was to not always load
the caja-html-sanitizer module. But as the module was still in the list
of dependencies, this lazy loading has never worked.

As most of the code is already migrated to not use jquery.sap.encoder
anymore (or will be migrated in the next weeks) it shouldn't be worth
looking into this again.

Therefore the unnecessary code is just removed.

Change-Id: I3cf46b4f442d6f231836ab0007d76d2d8c3c9b3f