[INTERNAL] XMLComposite: Increase coverage in aggregation templating
Note:
In order to improve the quality of the code in our XMLComposite I will
provide several small changes that are targeted to cover more code
inside our Unit tests. 
The aim here is to use the branch coverage scripts delivered by the core
team to identify in the earliest stage as possible dead code.

With this change I had the target to fully cover templateAggregations
method in the composite. Here we never tested single and unknown
aggregations.

Change-Id: If01eeb327896b02914a82c73c71b410653e99558