[FIX] sap.ui.unified.CalendarDateInterval: Sample now displays correctly on mobile
The issue was caused by setting big width (550px) on the
CalendarDateInterval that makes the control not fully visible on a small
mobile device (320px width). As the controls is designed to work with
fixed width, the width was reduced to the minimum - 320px. If no width
is set, the control may change its width with the content change.

Change-Id: I991ae022ed42d4c31752b2761bdb3513bd6eace0
BCP: 1880363674