[INTERNAL] jquery.sap.stubs QUnit: Reduce requests / split into chunks
In Firefox/Edge this test causes memory / performance issues because
of the high number of scripts / requests.

Therefore only the necessary modules will be unloaded to be loaded again
by the loader instead of completely reloading the loader.

Also, the tests can be executed in chunks, which will limit execution
to 50 tests to mitigate performance issues in those browsers.

Change-Id: Ie554319b38139aa4cee10f41d4060f112db052e2
BCP: 1880433534