[FIX] performance/trace/FESR: reenabled FESR activation
By refactoring the jquery.sap.trace dependencies out of Core modules
it was missed to reenable the activation of FESR somewhere else.
This caused a severe issue which made it impossible to activate FESR
at all. By moving the activation into a separate module required by
Core.js it can be decided at an early stage, equivalent to where it
was initially.

Change-Id: I1280accf3502e1f5fc51f3f025618b089e152c4f
BCP: 1880434071