[INTERNAL] XMLComposite: Remove templating from core
Note:
As discussed, this change is intended to reduce the complexity of the
XMLComposite and to be better prepared for future changes.
With this change we remove all the templating/preprocessing stuff from
the openui5 and move that to the runtime in an MDCXMLCompoiste.
This change is in align with another runtime change.

Change-Id: I9c58d68a63d0f9759f6ce1c7b6b867d13acc3587