[INTERNAL] IconPool: Lazy load core resource bundle
The core resource bundle is now only loaded when an icon is used.
This does not only prevent loading it when not required but also helps
when the language is changed dynamically so that the correct text is
used. However it does not update any existing texts nor reload the
bundle in case it has already been loaded and the language is changed.
So there is still the chance for texts with incorrect language in case
the language is changed dynamically.

Change-Id: Ia107ec480515752d3b0865e7a9ea3d680e32f32c
BCP: 1880004704