[FIX]sap.ui.rta: Disable actions in relevant container without stable id
It was possible to perform actions on controls with relevant container
not having a stable id (e.g. Rename group on SimpleForm when the
SimpleForm didn't have a stable ID). This created invalid changes and
even caused dumps in runtime.

This fix adds a check for the stable ID on the relevant container when
needed.

Change-Id: I48c3e16bc48b9558cba462ce6ff64339bf20cd2e
BCP: 1880415501
CR-Id: 002075125900003194082018
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>