[FIX]sap.ui.rta: Disable actions in relevant container without stable id
It was possible to perform actions on controls with relevant container
not having a stable id (e.g. Rename group on SimpleForm when the
SimpleForm didn't have a stable ID). This created invalid changes and
even caused dumps in runtime.

This fix adds a check for the stable ID on the relevant container when
needed.

Change-Id: I86f6fae3d9f0cf494e014aa3a9b218e85c4d1403
BCP: 1880415501
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>