[INTERNAL] sap.ui.dt control domRef in dt-metadata
Allow control domRef description in designtime metadata also for
controls that have a domRef, before it could only used by elements that
had not domRef themself. In addition allow it to be a selector and not
only a function.

Additional fix for VariantManagement designtime: Have an arbitrary
entry for the title domRef needed to know where the rename handle should
be created.

Change-Id: Ifb729f58b92e92d931bc88b5f39c3f6972cb073b