[INTERNAL] Replace jquery.sap.* leftovers in framework code base
This change replaces the jquery.sap.* leftovers which could not be
discovered automatically by the replacement script.

Note:
Not all leftovers are contained in this change.
There are still leftovers in the code base.
Mostly synchronous jQuery.sap.* calls, which need to be looked into
by the respective control developers.

Change-Id: Ib8772e5e8b03ae47aa8607339bfa9b9117075a55
JIRA: CPOUIFPHOENIXCORE-2587