[INTERNAL][FIX] Core: improve behavior in case of repeated execution
When the Core accidently was executed twice (e.g. in case of an async /
sync loading conflict), the second execution failed with a runtime error
as it tried to retrieve the facade of a facade.

The check for repeated execution now is done much earlier and returns
the facade created by the first execution.

This doesn't mean that double execution of the Core is now a supported
feature of UI5. Async / sync conflicts still need to be resolved, e.g.
by fully migrating to async APIs.

Change-Id: I4af86c81e06d96285dc87a239a0e63065adee862