[FIX] OverflowToolbar: Resize is now handled correctly when in Dialog
Problem: If the "OverflowToolbar" is in "Dialog", on iPhone 6/7/8
there is a bug in the layout when the orientation is changed
from portrait to landscape.

Root Cause: "doLayout" cache small values of the content, which results
to wrong calculations when the width of the "OverflowToolbar" gets bigger
when the orientation changes from portrait to landscape.

Solution: We reset and rerender the "OverflowToolbar" with the newly added
property "_bForceRerenderOnResize" when needed. In our case in the Dialog.

Change-Id: I75e50dcb7e1a45aea2aa9087b8e1cb4101969359
BCP: 1870031078