[INTERNAL] XMLComposite: Cloning of runtime composite
Note:
Taking all fact of the split and the problems in the cloning into
account we identified that for a runtime composite like the core
composites are there is no need and nothing to win from doing any
special handling on the cloning. The only control where we have to
investigate in cloning are that controls that use templating.
Thus we remove all the cloning respective changes here inh the core
composite and address the issues in cloning on the root problem, i.e.
inside the MDC composites.

Change-Id: I15bf1ec8556b997304d3b401874d30f621aeac1f