[INTERNAL]sap.ui.dt - Performance tests adjustments
The DT performance tests included the loading time of the modules
required to build and use DT (e.g. plugins). However, this loading is
not part of DT functionality and therefore should not be included in the
time measured by the tests, since changes in the loading (from core)
might have a detrimental effect on our tests even if we didn't change
anything from our side.

Change-Id: I050287f66bbc5d64592ac8de4dcfae1e39d533ff
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>