[INTERNAL] Table: Add TableUtils#isA (formerly TableUtils#isInstanceOf)
Add TableUtils.isA method as a wrapper for sap.ui.base.Object.isA, to
avoid additional dependencies to the base object. TableUtils is
anyway present almost everywhere.
This is now a replacement for the previously removed
TableUtils.isInstanceOf method.
If it is known that a variable is an instance of a UI5 object, still the
instance method #isA is used.

Change-Id: I46dcdea643b2660d44402f6a5d83e18412075688