[INTERNAL][FIX] Test Starter: IE11 issues (missing Promise)
This mainly reverts commit 41e67657d7a0b3cd1b21aacf2e4f20e7a152fa4f
as it made use of Promise before loading the actual Promise polyfill
which is required for IE11 and partly Safari (Promise#finally).

In case the TestRunner can't access the window.suite function after the
load event, it will wait for a CustomEvent which is triggered by the
createSuite.js once the window.suite function is available.
This should solve all issues in Firefox and IE.

The migrate.html has been adopted accordingly, but those changes
shouldn't be required as the relevant testsuites are old-style, with
window.suite available directly and not firing the event.

Change-Id: I2ff39e746cc34da04c8c7b973b8493f158708d1f
BCP: 1880474731