[FIX] sap.m.FacetFilter: update FacetPage on navigating
If a FacetFilter lists are binded to a model and a new list
is added on navigating the FacetPage was not update with
the new model data unless the FacetDialog is closed and
opened again.

Now when a naigation from a FilterPage is done to the
FacetPage the model is updated with the new data.

Change-Id: Id4b0c586cd1332b71989d601367b6e98dd40215a
BCP: 1880435856