[INTERNAL] sap.ui.support: fix double slash problem in RuleSetLoader
getAbsoluteUrl() worked incorrectly when it receives up-folder symbol at the
end of the relative URI, e.g.
getAbsoluteUrl("a/b/c/..") -> /a/b/c/
getAbsoluteUrl("a/b/c") -> /a/b/c

As a result of this, the final path to .supportrc files was calculated with double
slashes "/a/b/c//.supportrc" which cause an error on those webservers which are not
configured to squash multi-slashes into one (e.g. Jetty in our build infrastructure).

This fix makes sure that the returned valued never contains the ending slash.

Change-Id: I8fc27f356a5979f92125d29d02e5a74d6568cf5a