[FIX] sap.m.ComboBox: Events order fixed
There's a race condition where the focus jumps faster than the Combobox
events and breaks their order. In this case these are selectionChange and
change.
Because of the focusing, the "change" event is fired before the "selectionChange".
And as so, the code which is responsible for updating the properties
(it's in selectionChange handler) is not ready yet.

Unfortunately, it's not possible to write tests for that fix as the
reproduce path is quite specific and relies heavily on clicks and focusing
which cannot be emulated via sap.ui.test.qunit.triggerMouseEvent.

Change-Id: If86aa535d8798ec1121f26a5bdd8fd716edb9874
BCP: 1880480536
CR-Id: 002075125900003753042018