[FIX]sap.ui.rta - Evaluate AddOData Action after MetaModel is loaded
When evaluating if "Add Field" should be available for a control, it is
possible that the MetaModel was still not loaded for that control, which
makes the action unavailable. This fix adds an extra check for the
action after the MetaModel is loaded, ensuring that the action will be
there if all conditions are met.

BCP: 1880499137
CR-Id: 002075125900003525952018

Change-Id: I41aa84ae120eb169f7f764e936f7c9056973f5cd
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit e9d8dba745b53786b9db1b7b7ff9b336e7a6a021)