[FIX] Manifest: set component base URL
The component base URL was not correct set in the manifest.
This issue only occurs if the feature "enhanceWith" is used in the
manifest. When the property bundleUrlRelativeTo: "manifest" is set,
the manifest may use the base URL of the component to set the manifest
base URL.

Change-Id: If56874c705d681525d1eef7c4ee3a378ecf1dbed
CR-Id: 002075125900003500772018
BCP: 1880503434
(cherry picked from commit ca9f4106a84edc5e5a259e33ac89fa724081cea1)