[INTERNAL] Test Starter: register QUnit callbacks in correct order
In order to avoid the embedded sync loading of qunit-reporter.junit.js
(reporter) during the execution of qunit-junit.js (bridge), the new test
starter loads the reporter before loading the bridge. 

While this avoids one sync request and eval, it has the side effect that
bridge and reporter now register their QUnit callbacks in the wrong,
inverse order. As a consequence, test names are prefixed with the test
page name too late, only after their names have been recorded for the 
JUnit report.

With this change, the order of most callbacks is corrected again. To
achieve this, the test starter first creates a mock version of the
reporter's API, which only collects the registered callbacks. Then it
runs the bridge, then the reporter and finally applies the collected
callbacks to the real reporter API.

This ensures the right order of the callbacks for 'moduleStart',
'testStart' (and 'testDone'), but not of those for 'log'. However, the
two 'log' callbacks seem to be independent, so their order shouldn't
matter.

(cherry picked from commit 782b6dc60468a61f11ee0089cd03b6d8c2ffdd40)
Change-Id: Icb6790bb28c16fb08d1bdd61b826433131772c7c
CR-Id: 002075125900003596882018
BCP: n/a