[FIX] sap.m.PlanningCalendar: shows selectAll checkBox at the correct position
Previously the interval of the calendar was always placed at the end
of the infoToolbar content, which contains the calendar header,
the interval and the selectAll checkbox(if needed).

Now the interval is inserted to the right position considering
the selection mode and the row headers state of the calendar.

Change-Id: I257e515bd626a5cf5690742601cc5c91e973c95c
BCP: 1880487932
(cherry picked from commit 4d899d72559159841bd16dff86a1bcf507bf5bbb)
CR-Id: 002075125800004201092018