[FIX]sap.ui.rta - Change Handler error for instance-specific flexibility
When retrieving the change handler for a control with instance-specific
flexibility file there was an error where a function expected a control
and was receiving a selector. The change handler was then not found.

This change fixes the order of the return value of the function.

CR-Id: 002075125900003616962018
BCP: 1880468643

Change-Id: Iadf9369d0843ad8d7c98e3e99facb717ca6665e3

Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit ae2d300f3f4e1c1fb778316edecec534b36fe447)

Change-Id: I25b257a86c231e5f4eca73c96ee61f5d1de5f2f7