[FIX]sap.m - Table Designtime DomRef definition for items aggregation
Since a recent RTA fix where the aggregation overlays are rendered in
the correct order, it has been impossible to select the column overlays
in the table because the items aggregation overlay was on top of the
entire table.

This fix sets the correct DomRef to the items aggregation, not including
the table header and allowing the columns to be selected again.

Change-Id: I63732a5e86c602cdd41663a15e82ee68aaa2ac98
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>