[FIX]sap.ui.rta - Check for aggregation binding only for overlays
In the RTA BasePlugin there is a check if a stable element is in an
aggregation binding. This check was being performed even on elements
that did not have overlays, causing an exception on follow-up steps that
expected an overlay to be there.

This change restricts the check only to elements that have Overlays.

CR-Id: 002075125900003577042018
BCP: 1880500067

Change-Id: I2830adefc32b0509929e0034d0b0f51546e1c0b2
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit f80f3635055498cc8c43ec1f67f161f72a6a22d9)