[FIX] ObjectPageHeader - Fixed styles of navigationBar
The Problem: When in Fiori Launchpad, the navigationBar aggregation
has a visible navigation back icon in the left corner, although
the Fiori 2 adapter is trying to hide it.

The Root Cause: The root element of navigationBar aggregation -
sap.m.Bar has height of 0, but its children have height of 2rem.

The Solution: We set the root element display style to none.

Change-Id: I5e99e396693a86136669fe0ee0767b17839bd529
BCP: 1880457387