[FIX]sap.ui.dt - ScrollContainer Z-Index calculation fix
With the rewriting of the z-index calculation in overlays, the
calculation for the scroll container was missing when its position was
set.

This change adds an extra call to _setZIndex() to ensure that the
z-index of the scroll container is correct.

CR-Id: 002075125900003650972018
BCP: 1880499912

Change-Id: Ie456d33e41954b495640d53495595d33ac7f2034
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit b7bf7d39f3562666bdc2d5c776287a86b816962a)