[FIX] Aggregation Forwarding: also clone forwarded children
PS1: only the functional change itself

PS2: fixing the voter test errors:
- One error was caused by the unit test itself -> not a problem.
- The second one (in Input) was caused by the SuggestionPopup, albeit
internal, being added into the (public) "dependents" aggregation - which
means it is cloned, including the contained suggestion rows etc. When
then the suggestion rows are cloned because they have been forwarded:
boom. (duplicate ID)

PS3: when checking all controls that A) use forwarding and B) override
the clone method, two further issues were discovered:
- IconTabBar manually clones the internal IconTabHeader. Being in a
hidden aggregation it was not cloned by the framework, true. BUT
actually it was created in the clone's init method, so until now every
IconTabBar cloning created TWO headers! So removing the clone method
fixes a bug, anyway. The problem (which the clone method was supposed to
solve) is that the cloned header is not populated with cloned items
(because they are no longer in a public aggregation after forwarding -
no matter whether forwarding is manual or central). And this issue is
resolved with this change.
- Pretty much the same is the case for the items which are cloned by the
Select.

PS4: Implementation cleanup. Nice now.

PS5: accidental empty commit

PS6:
- ComboBox was just like Select
NotificationListBase: clones the overflow toolbar even though one was
already created in init of the clone. And all children. Now the children
will be cloned directly and added to the initial overflow toolbar of the
clone.
- NotifactionListGroup and NotificationListItem unnecessarily overwrote
clone()

PS7: address review concerns
- new hidden aggregation in Input, replacing dependents
- rename mAllAggregations to avoid confusion

PS10: rebase/merge

Change-Id: I5bf04f22b6ebc2dd1377415ddcfd3de3188a3fb1