[FIX] sap.m.FacetFilter: search with relative item binding works now.
Previously, if an FacetFilterItem used to be bound with a relative syntax
template: new sap.m.FacetFilterItem({
  text: { path: "modelname>" }
})

the search functionality didn't work.

THe issues was that paths like "" (this is how modelname> is resolved)
was not used to construct a sap.ui.model.Filter.

Fixes https://github.com/SAP/openui5/issues/2124
Closes https://github.com/SAP/openui5/pull/2124
BCP: 1880474806

Change-Id: I8c7b2444886604eb4072b3d742149bd095433b2c
CR-Id: 002075125900003662412018