[FIX] sap.m.ObjectAttribute: doesn't wrap in ObjectPageHeaderContent in IE
- for other uses of OA we need its parts to be inline-block,
otherwise the focus gets broken. In ObjectPage styles of the OA are
overwritten in order not to wrap when there is enough space, but in case
it wraps the focus will be broken.

- this change also fixes firing of press event only when it is taped
on the link part of the ObjectAttribute

BCP: 1880362200

Change-Id: Ia294041e99aaa44c02f1d1f16a01d6eb012f57ab
(cherry picked from commit 154fb0e7cfd113088bcf88bb28489767d50c9bdd)
CR-Id: 002075125900003736322018