[FIX]sap.ui.layout - Flexibility: Check actions based on form layout
Some Form layouts are still available even if they are not supported. To
ensure that RTA does not break in these cases, a check was added to set
the actions to "null" on form layouts that are not supported. Also a
supported and a non-supported layouts were added to the RTA test app for
future testing.

BCP: 1880536682
CR-Id: 002075125800002986842018

Change-Id: I84a8096d9a9b0fe755f583f7363b94923a3b971f
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit cea62422cec6680c6501c6c4211eff10054e2f21)