[INTERNAL] Core: protect Core against direct usage of its constructor
Only the framework must create an instances of the Core class during
bootstrap. Application code or other code must not call the constructor
but call sap.ui.getCore() instead to get access to the Core facade.

As long as globals are not switched off, the Core constructor is still
visible and apps might mistakenly call it. This change ensures that
for such subsequent calls, the same facade as for the first instance is
returned and that an error is written to the Log.

Change-Id: I9a23a03a34092a5dadbf4dcc5ba91c96dd788ce8
CR-Id: 002075125900003761922018
BCP: 1880526004