[INTERNAL] SDK: API Reference - Tree links for module children fixed
The problem:
Links to module children are not working and opening a not found page
when opened from the API Reference tree.
For example "sap/base/Log.Level"

The root cause:
Currently we escape module names by replacing the "/" character with
"." so we can use it as a valid URL parameter and we are replacing
it back with "/" when we detect a module in the URL parameter. This is
working for module's but is failing for module children like
"module:sap/base/Log.Level"

The solution:
Instead of converting module name's we could use URL encoding/decoding
to achieve escaping the reserved URL characters like "/" and restoring
them back which will be more robust

Change-Id: I0fa9458aeb91fcb1390157fdaab28fa97802087b