[FIX] Image: Fixed broken indicator in some browsers
The Problem: A broken image indicator(icon) is shown in some browsers,
when the "src" property is invalid.

The Root Cause: We are removing the "sapMNoImg" when we are not
supposed to. By doing so, we are making the "broken" indicator
icon - visible.

The Solution:
We are making the image visible(removing the "sapMNoImg" css class) only
and only if we have "alt" property set to a custom string value and
"decorative" set to false.

The "decorative" property is mandatory to be set to false, in order for
"alt" to take effect.

Change-Id: I2bfaeeaee81370ef1103ca5cc9420964cb52063f
BCP: 1880526262