[FIX] sap.m.TimePicker - when pattern contains single quotes, mask is correct now.
When pattern is like "HH:mm 'ч'" (contains single quote char(s)), the mask
was not correct- it used to include both single quotes.

The DateFormat, thought, knows how to "skip" the quotes (i.e. it knows
the content between the quotes is a 'static' or 'hardcoded' string).

As a result, the length of the formatted value differs from the one expected
by the mask and this used to generate wrong input - for example
"  17:34:17 ч." instead of "17:34:17 ч.".

If end user focuses, the value changes to " 1:34:17 'ч'."

Change-Id: Ibad7d048f17f87f2f0da82ad33014d838cc37e88
BCP: 1870331728