[INTERNAL] sap.m.Table: Fix test page
When the Growing button in the footer toolbar is toggled, the table's
item aggregation is bound. As the "key" property was missing in this case,
the list item would be rerendered and focus was lost when growing is
enabled. Hence the item aggregation binding is updated to have a key.
BCP: 1870333786

Change-Id: I3eeb5b0a85af62bd3e514f893ca4df25d4b8a191