[INTERNAL][FIX] Test Starter: make recent improvement work on IE11
The recent improvement of allowing tests to return promises to wait for
before starting QUnit, unfortunately made all tests fail on IE11 as
an "arguments" object ended up in Promise.all. IE11 throws an error
when an array method (other than slice) is executed on an "arguments"
object.

The fix in the requireP wrapper now clones "arguments" into a normal
array before resolving the promise with it.

Change-Id: I3da346863b2b67a592095d5fadc9202240c612c9