[FIX] AnchorBar: QUnit test failure is fixed
QUnit test, that checks the width of a MenuButton with long text, is
moved before the test, that sets a CSS class to the HTML tag, which applies
styles for smaller viewport (Phone view). This style results in smaller width
of the MenuButton and that is why it cannot exceed the max-width of 12rem.

Change-Id: I94a368a597554754225e727e25f61a49c775e15e
BCP: 1870331740