[INTERNAL] sap.ui.test.TestUtils: fixtures for DELETE/PATCH/POST
Now it is possible to add fixtures for non-GET requests. The key is
"METHOD URL" (e.g. "POST Articles") in this case. For GET requests
the method is optional, so that existing fixtures are not invalidated.

Only the MusicArtists sandbox had to be adjusted, because it added an
incomplete extension.

PS1: done
PS2: clean up
PS3: review comments

Change-Id: Ia2e215dfcab91d84a040f5520a6caf6d8404fda1