[FIX] XMLView: fix re-rendering of directly nested XMLViews
So far, an XMLView could not be a direct child of another XMLView.
The DOM preservation mechanism couldn't handle the re-rendering.
XMLViews could only be re-rendered safely when nested into normal
UI5 controls.

The root cause was that the handling of placeholders for the children of
an XMLView conflicted with the preservation algorithm. This has been
fixed by introducing a third ID-prefix for the so called "Temporary"
DOM. When an XMLView is re-rendered, its new, temporary DOM uses that
prefix for its ID. This allows the onAfterRendering of the XMLView to
differentiate between the new DOM of a child XMLView and the placeholder
for the same child view indicating where it should be placed.

Additionally, the content preservation step must not only move
'to-be-preserved' controls to the preserve area but also create a
placeholder in their old location iff the parent control is also
preserved.

Change-Id: I9464163b04b177933722a62e2be124a721f0cd85
BCP: 1880410030
BCP: 1880495226