[FIX] SDK: MessageStrip for experimental is now displaying links
The Problem: The MessageStrip in Index of Experimental APIs in the
DemoKit is not showing links.

The Root Cause: "enableFormattedText" property of MessageStrip
is not set to true, thus it's impossible to display link in the
MessageStrip`s text.

The Solution: We set the "enableFormattedText" property to true
and add <a> tag in the text so the the link in it is displayed
properly.

Also the link itself is changed, due to the fact that the old
no longer exists.

Change-Id: I6a01730eb1ddb6c7a27393f2ac9fc399374ca321
BCP: 1880519179